# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/gradebooks/{gradebookId}/recognition-tickets": {
      "get": {
        "operationId": "recognitionList",
        "parameters": [
          {
            "name": "gradebookId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecognitionTicketDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["recognition"],
        "security": [{ "bearer": [] }, { "cookie": [] }]
      },
      "post": {
        "operationId": "recognitionUpload",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "gradebookId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image", "componentId"],
                "properties": {
                  "image": { "type": "string", "format": "binary" },
                  "componentId": { "type": "integer", "minimum": 1 },
                  "declaredRows": {
                    "type": "integer",
                    "minimum": 1,
                    "deprecated": true,
                    "description": "Deprecated và bị bỏ qua: hệ thống tự chốt danh sách lớp khi tạo phiếu (ADR-0015)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionReceiptDto"
                }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "413": { "description": "Image exceeds upload limit" },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["recognition"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{gradebookId}/recognition-tickets/{ticketId}": {
      "get": {
        "operationId": "recognitionDetail",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "gradebookId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionTicketDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["recognition"],
        "security": [{ "bearer": [] }, { "cookie": [] }]
      }
    },
...
  },
  "info": {
    "title": "Quản lý điểm API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "cookie": { "type": "apiKey", "in": "cookie", "name": "qld_session" },
      "csrf": { "type": "apiKey", "in": "header", "name": "x-csrf-token" }
    },
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "object", "nullable": true },
          "requestId": { "type": "string" }
        },
        "required": ["code", "message", "details", "requestId"]
      },
...
      "RecognitionTicketDto": {
        "type": "object",
        "properties": {
          "ticketId": { "type": "string" },
          "gradebookId": { "type": "number" },
          "componentId": { "type": "number" },
          "componentName": { "type": "string" },
          "declaredRows": { "type": "number" },
          "detectedRows": { "type": "number", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["DANG_XU_LY", "CHO_DOI_CHIEU", "DA_DUYET", "LOI"]
          },
          "errorCode": { "type": "string", "nullable": true },
          "modelVersion": { "type": "string", "nullable": true },
          "version": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "greenRows": {
            "type": "number",
            "description": "Số dòng mức Xanh (mức cuối)"
          },
          "yellowRows": {
            "type": "number",
            "description": "Số dòng mức Vàng (mức cuối)"
          },
          "redRows": {
            "type": "number",
            "description": "Số dòng mức Đỏ (mức cuối)"
          }
        },
        "required": [
          "ticketId",
          "gradebookId",
          "componentId",
          "componentName",
          "declaredRows",
          "detectedRows",
          "status",
          "errorCode",
          "modelVersion",
          "version",
          "createdAt",
          "greenRows",
          "yellowRows",
          "redRows"
        ]
      },
      "RecognitionEvidenceRowDto": {
        "type": "object",
        "properties": {
          "rowId": { "type": "string" },
          "order": { "type": "number", "description": "Vị trí dòng trên ảnh" },
          "stt": {
            "type": "number",
            "nullable": true,
            "description": "STT hệ thống trong danh sách lớp đã chốt của phiếu"
          },
          "sttOnPaper": {
            "type": "number",
            "nullable": true,
            "description": "STT in trên giấy đọc được"
          },
          "studentId": { "type": "number" },
          "studentName": {
            "type": "string",
            "description": "Họ tên học sinh theo danh sách lớp đã chốt"
          },
          "nameRead": {
            "type": "string",
            "nullable": true,
            "description": "Họ tên máy đọc được trên giấy"
          },
          "matchConfidence": {
            "type": "string",
            "nullable": true,
            "description": "Độ tin cậy ghép dòng với học sinh, 0..1"
          },
          "matchNote": { "type": "string", "nullable": true },
          "numericRaw": { "type": "string", "nullable": true },
          "numericValue": { "type": "string", "nullable": true },
          "numericConfidence": { "type": "string", "nullable": true },
          "writtenRaw": { "type": "string", "nullable": true },
          "writtenValue": { "type": "string", "nullable": true },
          "writtenConfidence": { "type": "string", "nullable": true },
          "comparison": {
            "type": "string",
            "enum": ["KHOP", "LECH", "MOT_KENH", "KHONG_DOC_DUOC"]
          },
          "reviewLevel": { "type": "string", "enum": ["XANH", "VANG", "DO"] },
          "finalValue": { "type": "string", "nullable": true },
          "numericCropUrl": { "type": "string", "nullable": true },
          "writtenCropUrl": { "type": "string", "nullable": true },
          "nameCropUrl": {
            "type": "string",
            "nullable": true,
            "description": "Ảnh ô họ tên trên giấy (URL ký 300 giây)"
          }
        },
        "required": [
          "rowId",
          "order",
          "stt",
          "sttOnPaper",
          "studentId",
          "studentName",
          "nameRead",
          "matchConfidence",
          "matchNote",
          "numericRaw",
          "numericValue",
          "numericConfidence",
          "writtenRaw",
          "writtenValue",
          "writtenConfidence",
          "comparison",
          "reviewLevel",
          "finalValue",
          "numericCropUrl",
          "writtenCropUrl",
          "nameCropUrl"
        ]
      },
      "RecognitionTicketDetailDto": {
        "type": "object",
        "properties": {
          "ticketId": { "type": "string" },
          "gradebookId": { "type": "number" },
          "componentId": { "type": "number" },
          "componentName": { "type": "string" },
          "declaredRows": { "type": "number" },
          "detectedRows": { "type": "number", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["DANG_XU_LY", "CHO_DOI_CHIEU", "DA_DUYET", "LOI"]
          },
          "errorCode": { "type": "string", "nullable": true },
          "modelVersion": { "type": "string", "nullable": true },
          "version": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "greenRows": {
            "type": "number",
            "description": "Số dòng mức Xanh (mức cuối)"
          },
          "yellowRows": {
            "type": "number",
            "description": "Số dòng mức Vàng (mức cuối)"
          },
          "redRows": {
            "type": "number",
            "description": "Số dòng mức Đỏ (mức cuối)"
          },
          "sourceImageUrl": { "type": "string" },
          "imageUrlExpiresInSeconds": { "type": "number", "example": 300 },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecognitionEvidenceRowDto"
            }
          }
        },
        "required": [
          "ticketId",
          "gradebookId",
          "componentId",
          "componentName",
          "declaredRows",
          "detectedRows",
          "status",
          "errorCode",
          "modelVersion",
          "version",
          "createdAt",
          "greenRows",
          "yellowRows",
          "redRows",
          "sourceImageUrl",
          "imageUrlExpiresInSeconds",
          "rows"
        ]
      },
      "RecognitionReceiptDto": {
        "type": "object",
        "properties": {
          "ticketId": {
            "type": "string",
            "description": "bigint recognition ticket ID"
          },
          "jobId": { "type": "string" },
          "status": { "type": "string", "enum": ["DANG_XU_LY"] }
        },
        "required": ["ticketId", "jobId", "status"]
      },
...
    }
  }
}